+++ screenshots/square_map_with_resources.png
