+++ screenshots/square_map_with_resources_and_players.png
